# openapi.yaml
openapi: 3.0.3
info:
  title: Insurance Brokerage API
  description: |
    API for the Insurance Brokerage application that allows management of clients, documents, and user authentication.
  version: 1.0.0
  contact:
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.insurance-brokerage.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
    Client:
      type: object
      required:
        - customer_type
        - product
        - insurance_provider
        - client_name
        - mobile_no
      properties:
        id:
          type: string
        introducer_code:
          type: string
        customer_type:
          type: string
        product:
          type: string
        policy_:
          type: string
        insurance_provider:
          type: string
        branch:
          type: string
        client_name:
          type: string
        street1:
          type: string
        street2:
          type: string
        city:
          type: string
        district:
          type: string
        province:
          type: string
        telephone:
          type: string
        mobile_no:
          type: string
        contact_person:
          type: string
        email:
          type: string
          format: email
        social_media:
          type: string
        nic_proof:
          type: string
        dob_proof:
          type: string
        business_registration:
          type: string
        svat_proof:
          type: string
        vat_proof:
          type: string
        coverage_proof:
          type: string
        sum_insured_proof:
          type: string
        policy_fee_invoice:
          type: string
        vat_fee_debit_note:
          type: string
        payment_receipt_proof:
          type: string
        policy_type:
          type: string
        policy_no:
          type: string
        policy_period_from:
          type: string
          format: date
        policy_period_to:
          type: string
          format: date
        coverage:
          type: string
        sum_insured:
          type: number
        basic_premium:
          type: number
        srcc_premium:
          type: number
        tc_premium:
          type: number
        net_premium:
          type: number
        stamp_duty:
          type: number
        admin_fees:
          type: number
        road_safety_fee:
          type: number
        policy_fee:
          type: number
        vat_fee:
          type: number
        total_invoice:
          type: number
        debit_note:
          type: string
        payment_receipt:
          type: string
        commission_type:
          type: string
        commission_basic:
          type: number
        commission_srcc:
          type: number
        commission_tc:
          type: number
        policies:
          type: number
    ClientResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Client'
    ClientsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Client'
        totalCount:
          type: integer
          description: Total number of clients matching the request (for pagination)
    CreateClientResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
    UpdateResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /api/clients:
    get:
      tags:
        - Clients
      summary: Get all clients
      description: Get paginated list of all clients with optional search functionality
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of clients to return (default 100)
        - in: query
          name: offset
          schema:
            type: integer
          description: Number of clients to skip (default 0)
        - in: query
          name: search
          schema:
            type: string
          description: Search term to filter clients by name or mobile number
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientsResponse'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Clients
      summary: Create a new client
      description: Create a new insurance client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '201':
          description: Client created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClientResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/clients/{id}:
    get:
      tags:
        - Clients
      summary: Get client by ID
      description: Get a specific client by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientResponse'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Clients
      summary: Update client
      description: Update a client's information
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Client ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Client updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Clients
      summary: Delete client
      description: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Client ID
      responses:
        '200':
          description: Client deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResponse'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/clients/search:
    post:
      tags:
        - Clients
      summary: Search clients
      description: Search for clients based on criteria with pagination support
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of clients to return (default 100)
        - in: query
          name: offset
          schema:
            type: integer
          description: Number of clients to skip (default 0)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientsResponse'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/clients/import-csv:
    post:
      tags:
        - Clients
      summary: Import clients from CSV
      description: |
        Import multiple clients from a CSV file. 
        
        For large files (500+ rows), the response will be streamed with progress updates during processing. 
        The data is processed in batches of 25 records to improve performance and reduce database load.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file containing client data. Must include required fields - customer_type, product, insurance_provider, client_name, mobile_no
      responses:
        '200':
          description: |
            Clients imported successfully. 
            
            For large imports, response is streamed as chunked transfer encoding with progress updates.
            Each chunk is a JSON object containing current progress.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  count:
                    type: integer
                    description: Number of clients successfully imported
                  totalCount:
                    type: integer
                    description: Total number of records to process
                  processedCount:
                    type: integer
                    description: Number of records processed so far
                  progress:
                    type: integer
                    description: Percentage of import completed (0-100)
                  ids:
                    type: array
                    items:
                      type: string
                    description: IDs of the created clients
        '400':
          description: Invalid request - No file uploaded or CSV missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...